[#162592036 - moved maxLengthMap into general utils
