+++ app/javascript/components/applications/application_form/utils.js
@@ -4,22 +4,3 @@ export const mailingAddressFieldMap = {
   state: 'mailing_state',
   zip: 'mailing_zip_code'
 }
-
-export const maxLengthMap = {
-  first_name: 50,
-  middle_name: 50,
-  last_name: 50,
-  email: 50,
-  password: 50,
-  phone: 40,
-  address: 75,
-  city: 75,
-  state: 2,
-  zip: 15,
-  alternate_contact_type_other: 200,
-  agency_name: 250,
-  certificate_number: 50,
-  gender_other: 100,
-  sexual_orientation_other: 100,
-  income: 16
-}
