+++ app/javascript/utils/formUtils.js
@@ -40,6 +40,28 @@ const submitErrors = (formApi) => {
   return formApi.submits > 0 ? formApi.errors : mapValues(formApi.errors, () => null)
 }
 
+export const maxLengthMap = {
+  first_name: 40,
+  middle_name: 20,
+  last_name: 40,
+  day: 2,
+  month: 2,
+  year: 4,
+  email: 40,
+  password: 50,
+  phone: 40,
+  address: 75,
+  city: 75,
+  state: 2,
+  zip: 15,
+  alternate_contact_type_other: 200,
+  agency_name: 250,
+  certificate_number: 50,
+  gender_other: 100,
+  sexual_orientation_other: 100,
+  income: 16
+}
+
 export default {
   toOption,
   toOptions,
